+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change log
 
+- 0.8.0.pre - Auto-detect variable type.
+- 0.7.0.pre - Display exclusion rule in the report
 - 0.6.2.pre - Bugfix for 1D array of numbers scenario.
 - 0.6.1.pre - Bugfix on the converters when used with rule exclusion.
 - 0.6.0.pre - Reverted redesign.